+++ NEWS.md
@@ -1,2 +1,10 @@
+# argonDash 0.2.0.9000
+## Minor improvement
+- fix #3 thanks @noahbishop: when vertical is FALSE in the `argonDashSidebar`,
+navbar elements are horizontally displayed, as expected
+
+## Bug fix
+- fix #2: `argonDropdown` correctly render in the navbar
+
 # argonDash 0.1.0
 - first release